+++ src/types/Patterns.test.ts
@@ -93,7 +93,7 @@ const TESTDATA = {
       'www.bmw.de',
       'BMW.COM',
       '4chan.org',
-      'portal.env.example.catena-x.net',
+      'portal.env.example.railway-x.net',
       'www.5555.site', //valid?
       '1.2.3', //valid?
     ],
@@ -118,8 +118,8 @@ const TESTDATA = {
       'https://www.bmw.com/index.html',
       'https://hostname.domain:12345',
       'https://hostname.domain:12345/endwith/slash/',
-      'https://portal.env.example.catena-x.net/assets/swagger/index.html',
-      'https://portal.env.example.catena-x.net/assets/swagger/index.html?param=13&arg=value',
+      'https://portal.env.example.railway-x.net/assets/swagger/index.html',
+      'https://portal.env.example.railway-x.net/assets/swagger/index.html?param=13&arg=value',
     ],
     invalid: [
       '',